Ignore broken test on 2.1.1 and 2.1.2
Closes #26

It's a Ruby issue that they fixed but hasn't been released yet.

Example  here: https://gist.github.com/JoshCheek/8aee8ef4a5f8593db351
Reported here: https://bugs.ruby-lang.org/issues/9820
Fixed    here: https://github.com/ruby/ruby/commit/3771a370ad64aae87f751751e80d52d02a1735a9

+++ spec/hard_core_ensure_spec.rb
@@ -56,18 +56,21 @@ describe SeeingIsBelieving::HardCoreEnsure do
   end
 
   it 'invokes the code even if an interrupt is sent and interrupts are set to ignore' do
-    channel = IChannel.new Marshal
-    pid = fork do
-      old_handler = trap 'INT', 'IGNORE'
-      result = call code: -> { sleep 0.1; 'code result' }, ensure: -> { channel.put "ensure invoked" }
-      channel.put result
-      trap 'INT', old_handler
+    pending 'need parser to work for 2.2' do
+      fail if RUBY_VERSION == '2.1.1' || RUBY_VERSION == '2.1.2'
+      channel = IChannel.new Marshal
+      pid = fork do
+        old_handler = trap 'INT', 'IGNORE'
+        result = call code: -> { sleep 0.1; 'code result' }, ensure: -> { channel.put "ensure invoked" }
+        channel.put result
+        trap 'INT', old_handler
+      end
+      sleep 0.05
+      Process.kill 'INT', pid
+      Process.wait pid
+      channel.get.should == "ensure invoked"
+      channel.get.should == 'code result'
+      channel.should_not be_readable
     end
-    sleep 0.05
-    Process.kill 'INT', pid
-    Process.wait pid
-    channel.get.should == "ensure invoked"
-    channel.get.should == 'code result'
-    channel.should_not be_readable
   end
 end
